Update library to no-std
The library is intended to be used in contracts/WASM environments.

For this reason, it should be no-std.

Resolves #9
