+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add simplified default type [#6](https://github.com/dusk-network/dusk-hamt/issues/6)
+- Update library to no-std [#9](https://github.com/dusk-network/dusk-hamt/issues/9)
 
 ## [0.2.0]
